pyproject.toml
[tool.poetry]
name = "expire"
version = "0.1.0"
description = "A tool for expiring old backups."
authors = ["Ben Steinberg <bsteinberg@law.harvard.edu>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.3"
croniter = "^1.3.7"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest-click = "^1.1.0"
pyproject-flake8 = "^5.0.4"

[tool.poetry.scripts]
expire = 'expire:expire'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
extend-ignore = ["W191", "E101"]
